+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 # 1.1.0
 - Fixed a problem with automcompletion for the weaponHudInterface structure
+- Added x and y properties to weaponHudInterface structure
 - Added landing property to biped structure
-- Added walkingDirection to biped structure
 
 # 1.0.0
 - Initial release